Make all tests pass
Refactored to use BigInteger.

+++ KParkov.Distributions/PermutationCount.cs
@@ -1,10 +1,12 @@
+using System.Numerics;
+
 namespace KParkov.Distributions;
 
 public readonly record struct PermutationCount
 {
     public required int Value { get; init; }
-    public required long Permutations { get; init; }
+    public required BigInteger Permutations { get; init; }
     public required double Ratio { get; init; }
-    public required long AtLeast { get; init; }
-    public required long AtMost { get; init; }
+    public required BigInteger AtLeast { get; init; }
+    public required BigInteger AtMost { get; init; }
 }
